translation-system/requirements.txt
# Core dependencies
deep-translator==1.11.4
python-dotenv==1.0.0
paramiko==3.3.1

# File processing
chardet>=3.0.4
beautifulsoup4==4.12.2
markdown==3.5.1

# Logging and monitoring
colorlog==6.7.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development tools
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Security
cryptography==41.0.7

# Optional: Advanced translation services
openai==0.28.1

# System monitoring
psutil==5.9.6

# Async support
aiofiles==23.2.1
asyncio-throttle==1.0.2
